back_axf/requirement.txt
Django==2.0.7
django-cors-headers==2.5.3
django-filter==2.1.0
django-redis==4.10.0
djangorestframework==3.9.3
PyMySQL==0.9.3
redis==3.2.1
